Implemented mutex tracking for the excel_upload function offered by the file management system

+++ initializers/post_PDF_dataExtraction_tasks.py
@@ -8,6 +8,7 @@ import os
 from data_stores.AzureBlobObjects import AzureBlobObjects as abo
 from data_stores.AzureTableObjects import AzureTableObjects as ato
 import config
+import platform
 
 commandLineArguments = sys.argv
 pdfFilepath = None
@@ -19,13 +20,14 @@ if len(commandLineArguments) == 6:
     mutexKey = commandLineArguments[4]
     chapterNumber = int(commandLineArguments[5])
 
-os.remove(pdfFilepath)
-ato.edit_entity(chapterNumber, mutexKey, newRecordStatus=config.RecordStatus.uploadingGeneratedDocuments)
+if platform.system() != 'Windows': # had issues with Windows (at least the Browns laptop) where the file was still 'in-use' even after the response was created. Shouldn't be an issue in deployment because we are using an Azure linux app service.
+    os.remove(pdfFilepath)
 abo.upload_blob_file(excelPath,config.generatedExcel_container_name) # upload generated excel to azure blob
 abo.upload_blob_file(dictPath,config.generatedDict_container_name) # upload dictionary pickle to azure blob
 ato.edit_entity(chapterNumber, mutexKey, newRecordStatus=config.RecordStatus.uploadedGeneratedDocuments, newRecordState=config.RecordState.pdfUploaded)
 ato.release_mutex(chapterNumber, mutexKey)
-os.remove(excelPath)
-os.remove(dictPath)           
+if platform.system() != 'Windows': # had issues with Windows (at least the Browns laptop) where the file was still 'in-use' even after the response was created. Shouldn't be an issue in deployment because we are using an Azure linux app service.
+    os.remove(excelPath)
+    os.remove(dictPath)           
 
 print("Data extracted from tariff pdfs and saved as excel (and text data dictionary pickle) for review.")
